pyproject.toml
[tool.poetry]
package-mode = false
name = "llm-langfuse-langchain-demo"
version = "0.1.0"
description = "A project using transformers, langchain and langfuse for language model applications"
authors = ["Bartosz Kopytek <bkopytek@gmail.com.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
torch = "^2.0.0"
transformers = "^4.30.0"
langfuse = "^3.3.4"
langchain = "^0.3.27"
langchain-community = "^0.3.29"
accelerate = "^1.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"
python-dotenv = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true